car code complete, untested
